refactor(core): split tests in happy path and exception handling
